+++ src/lib.rs
@@ -4,10 +4,12 @@ pub mod geojson_utils;
 pub mod jittered_hexagonal_sampling;
 pub mod projection;
 pub mod main_functions;
-pub mod shared_buffer;
 
 #[cfg(not(target_arch = "wasm32"))]
 pub mod requests;
 
 #[cfg(target_arch = "wasm32")]
 pub mod requests_wasm;
+
+#[cfg(target_arch = "wasm32")]
+pub mod shared_buffer;
